requirements.txt
notebook
numpy
pandas
web3
py-solc-x
# psycopg2       # windows implementaion
psycopg2-binary # linux and mac implemenation
flask
flask_jwt_extended